# core/grantforge-web/src/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "PageResultTenantResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/TenantResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "items", "page", "size", "total" ],
        "type" : "object"
      },
...
      "TenantCreateRequest" : {
        "properties" : {
          "adminDisplayName" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          },
          "adminPassword" : {
            "maxLength" : 1024,
            "minLength" : 0,
            "type" : "string"
          },
          "adminUsername" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "code" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "adminPassword", "adminUsername", "code", "name" ],
        "type" : "object"
      },
      "TenantResponse" : {
        "properties" : {
          "accounts" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "platform" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "ACTIVE", "SUSPENDED" ],
            "type" : "string"
          }
        },
        "required" : [ "accounts", "code", "createdAt", "id", "name", "platform", "status" ],
        "type" : "object"
      },
      "TenantUpdateRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "Fine-grained, plugin-based authorization platform",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/license/mit"
    },
    "title" : "GrantForge API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/tenants" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResultTenantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tenant-controller" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "tenant-controller" ]
      }
    },
    "/api/v1/tenants/{id}" : {
      "get" : {
        "operationId" : "find",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tenant-controller" ]
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tenant-controller" ]
      }
    },
    "/api/v1/tenants/{id}/activate" : {
      "post" : {
        "operationId" : "activate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tenant-controller" ]
      }
    },
    "/api/v1/tenants/{id}/suspend" : {
      "post" : {
        "operationId" : "suspend",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tenant-controller" ]
      }
    }
  },
  "servers" : [ {
    "url" : "/"
  } ]
}